+++ src/main/java/net/sirplop/aetherworks/effect/MoonfireEffect.java
@@ -5,6 +5,7 @@ import net.minecraft.world.damagesource.DamageSource;
 import net.minecraft.world.effect.MobEffect;
 import net.minecraft.world.effect.MobEffectCategory;
 import net.minecraft.world.entity.LivingEntity;
+import net.sirplop.aetherworks.AWConfig;
 import net.sirplop.aetherworks.Aetherworks;
 import net.sirplop.aetherworks.damage.DamageMoonEmber;
 import net.sirplop.aetherworks.datagen.AWDamageTypes;
@@ -19,7 +20,7 @@ public class MoonfireEffect extends MobEffect {
         if (!pLivingEntity.level().isClientSide()) {
             DamageSource source = new DamageMoonEmber(pLivingEntity.level().registryAccess().registry(Registries.DAMAGE_TYPE).get()
                     .getHolderOrThrow(AWDamageTypes.MOON_EMBER_KEY), null, null);
-            pLivingEntity.hurt(source, 4.0F);
+            pLivingEntity.hurt(source, AWConfig.MOONGAZE_STRENGTH.get().floatValue());
         }
 
         super.applyEffectTick(pLivingEntity, pAmplifier);