+++ src/main/java/net/sirplop/aetherworks/item/tool/SculkAxe.java
@@ -50,11 +50,11 @@ public class SculkAxe extends AOEEmberDiggerItem{
     @Override
     public @NotNull InteractionResult useOn(@NotNull UseOnContext context) {
         InteractionResult result = super.useOn(context);
-        if (!(context.getLevel() instanceof ServerLevel) || context.getLevel().isClientSide || AWConfig.getConfigSet(AWConfig.Tool.ENDER_AXE).isEmpty())
+        if (!(context.getLevel() instanceof ServerLevel) || context.getLevel().isClientSide)
             return result;
 
         if (context.getPlayer() == null
-                || !AWConfig.getConfigSet(AWConfig.Tool.SCULK_AXE).contains(context.getLevel().getBlockState(context.getClickedPos()).getBlock())
+                || !Utils.blockHasTag(context.getLevel().getBlockState(context.getClickedPos()), AWBlockTags.AOSA_ALLOWED)
         )
             return result;
         if (result == InteractionResult.PASS && context.getHand() == InteractionHand.MAIN_HAND